Use elm on $PATH rather than bundled version
As elm-format is currently in alpha, it can get out of date quickly.
Rather than rely on the packaged elm-format, this commit changes the
on-save script to look on the user's PATH for elm-format. This way we
can manage this dependency externally, rather than require a commit
every time elm-format changes.

Also update the README